Don't mount rack app at application level
This causes Rails middleware like ActiveRecord::Migration::CheckPending
to get in the way of LSP requests. We should surface migration errors
in the LSP instead of having features just stop working.

+++ lib/ruby_lsp_rails/rack_app.rb
@@ -8,10 +8,17 @@ module RubyLsp
 
       BASE_PATH = "/ruby_lsp_rails/"
 
+      sig { params(app: T.untyped).void }
+      def initialize(app)
+        @app = T.let(app, T.untyped)
+      end
+
       sig { params(env: T::Hash[T.untyped, T.untyped]).returns(T::Array[T.untyped]) }
       def call(env)
         request = ActionDispatch::Request.new(env)
         path = request.path
+
+        return @app.call(env) unless path.starts_with?(BASE_PATH)
 
         route, argument = path.delete_prefix(BASE_PATH).split("/")
